Added Basic Field finder front endpoint
This commit add a basic field finder page to the front end app.  the
field finder fetches the field statistics and dynamically updates the
returned fields as the field name and table name fields are updated.
It presents the returned values in an accordion. When the accordion
for a specific field is opened, the field stats are fetched from the
server and displayed to the user.  Still need to update the url query
parameters so that pages can be shared, and a check box array needs to
be added to filter returned project types.