Streamlining production with cherrypy
This commit adds a run_cherrypy script that serves the application in
a way that emulates our production environment.  Small tweaks to
several of the file to work with this set-up (which could still be
improved). The react app route in the flask api has been updated to
include a catch all that will pass unmatched paths to the front end to
be handled by react router.