Better defaults for react query and default export for pages
This commit changes all of the functions used to index pages to be the
default export to facilitate code splitting.  Splitting wasn't
implemented as it didn't reduce main bundle size significantly but did
produce a lot more files. It also includes better defaults for react
query (which is way to aggressive for data that never changes), and
also adds source to the dependency array for project list table
columns.  The links need to change when the source changes.